Basic Interface Commands
This commit focuses on fleshing out the basic interface commands via keyboard input.  After importing keyboard, I am able to monitor keyboard inputs.  This allows me to create a "kill switch" for the program, which will later develop into a toggle.

I was also able to create command macros.  These will be the bread and butter for the interface, as the AI model will call upon these functions to make decisions.  The macros remove the hard work of getting the AI to figure out how to navigate the menu.  This lets me tune the AI to a more battle strategy focus.