HP Value Correction Algorithm
Designed a brute-force algorithm to correct impossible / incorrect HP values based on the information available.

If there are impossible HP states (health bar is red but HP is greater than 0 for example), then the function will correct this.  If the current HP is greater than the max HP, the algorithm will also try to correct this.

Also defined a couple extra functions that will be coloured in later on:
- getHPstate will determine the state of HP based on pixel colour
     - If HP bar is red, HP = 0
     - if HP bar is yellow, HP <= 25% max HP

getHPvalues returns an array containing 2 arrays; current and max HP values, listed as numerators and denominators.