Main Control Script
Created a main control script with a flow to allow the learning model's "training mode" to be toggled.  Also added a thread to create a KillSwitch when pressing the esc key.  This thread is a barebones implementation, but will make sure to save the current learning and halt keyboard inputs from the script before exiting all processes.